+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Link to official StochSD forum inside StochSD.
 - Better support for PWA StochSD
 - Pasting two columns from speadsheet application into converter dialog works.
+- Code-highlighting with CodeMirror for ConverterDialog.
 
 ## Fixed
 - Ghosted Converter Bugs